First pass Mission generation  + Generate 8x8 map of nodes  - Remove mission database/smithy-ing  - Remove HallMission and related types  = Reduce GameIdType to u64

+++ crates/shared-net/src/types.rs
@@ -1,7 +1,7 @@
 pub type UserIdType = u128;
 pub type PasswordType = u128;
 pub type AuthType = u128;
-pub type GameIdType = u128;
+pub type GameIdType = u64;
 pub type PartType = u64;
 pub type SeedType = u64;
 pub type TimestampType = u64;
